[common] fixup spmatrix gnuplot

+++ dune/stuff/common/print.hh
@@ -492,9 +492,10 @@ template< class Matrix >
 void matrixToGnuplotStream(const Matrix& matrix, std::ostream& stream) {
   unsigned long nz = 0;
 
-  for (size_t row = 0; row < matrix.rows(); ++row)
+  // don't try to be clever and use an unsigned counter here
+  for (int row = 0; row < matrix.rows(); ++row)
   {
-    for (size_t col = 0; col < matrix.cols(); ++col)
+    for (int col = 0; col < matrix.cols(); ++col)
     {
       if ( matrix.find(row, col) )
         stream << row << "\t" << col << "\t" << matrix(row, col) << std::endl;
